Updated Prime Generation
More specifically, I removed that 1024 bits is as secure as you can get, because now that is not the case. I also changed the prime check so that it would regenerate new primes if there was an error. 

Note : I don't really know Go so there may be some syntax errors.