Fix sort testcase input sorted unintentionally
A slice is a references to its underlying array. If a slice is
passed as an argument and modified in a function, the original
slice is also modified.

As getSortedVersion and other sorting functions modify a given slice,
testcase inputs are unintentionally sorted beforehand.

To avoid that, passes a cloned slices to these functions.